% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_conll.R
\name{dep_conll}
\alias{dep_conll}
\title{Dependency Tree CoNLL}
\usage{
dep_conll(x)
}
\arguments{
\item{x}{List of YAP output}
}
\value{
tibble of CoNLL dependency tree
}
\description{
Dependency Tree CoNLL
}
\examples{
text <- "גנן גידל דגן בגן"
token <- "YourAPIToken"
yap_list <- yap(text, token)
dep_conll(yap_list)
}
