% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_lattice.R
\name{dep_lattice}
\alias{dep_lattice}
\title{Dependency Tree Lattice}
\usage{
dep_lattice(x)
}
\arguments{
\item{x}{List of YAP output}
}
\value{
tibble of Lattice dependency tree
}
\description{
Dependency Tree Lattice
}
\examples{
text <- "גנן גידל דגן בגן"
token <- "YourAPIToken"
yap_list <- yap(text, token)
dep_lattice(yap_list)
}
